wanbuffer/test/wanbuffer/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
et-xmlfile==1.1.0
idna==3.7
numpy==1.24.4
openpyxl==3.1.3
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
XlsxWriter==3.2.0
